feat: :sparkles: computeRate 方法 （写到一半，边界判断还没完善，测试用例 没跑完）

+++ src/utils/adapter.ts
@@ -15,7 +15,7 @@ import Decimal from 'decimal.js'
   })
  */
 export const defaultDecimalConfig: Decimal.Config = {
-  precision: 10, // 暂时使用10位精度，可根据需求灵活调整
+  precision: 16, // 暂时使用16位精度，可根据需求灵活调整
   rounding: 4, // 使用标准四舍五入 5进位 4舍去
   toExpNeg: -7,
   toExpPos: 21,
